+++ components/search/bills/BillSearch.tsx
@@ -18,7 +18,6 @@ import { SearchErrorBoundary } from "../SearchErrorBoundary"
 import { useRouting } from "../useRouting"
 import { BillHit } from "./BillHit"
 import { useBillRefinements } from "./useBillRefinements"
-import { useBillHierarchicalMenu } from "./useBillHierarchicalMenu"
 import { SortBy, SortByWithConfigurationItem } from "../SortBy"
 import { getServerConfig } from "../common"
 import { useBillSort } from "./useBillSort"
@@ -103,7 +102,6 @@ const Layout: FC<
   React.PropsWithChildren<{ items: SortByWithConfigurationItem[] }>
 > = ({ items }) => {
   const refinements = useBillRefinements()
-  const hierarchicalMenu = useBillHierarchicalMenu()
   const status = useSearchStatus()
 
   const { t } = useTranslation("billSearch")
@@ -115,14 +113,12 @@ const Layout: FC<
       </Row>
       <Row>
         <Col xs={3} lg={3}>
-          {hierarchicalMenu.options}
           {refinements.options}
         </Col>
         <Col className="d-flex flex-column">
           <RefinementRow>
             <ResultCount className="flex-grow-1 m-1" />
             <SortBy items={items} />
-            {hierarchicalMenu.show}
             {refinements.show}
           </RefinementRow>
           <CurrentRefinements
